api/requirements.txt
annotated-types==0.6.0
anyio==3.7.1
asgiref==3.7.2
bcrypt==4.0.1
click==8.1.7
colorama==0.4.6
Django==4.2.6
dnspython==2.4.2
ecdsa==0.18.0
fastapi==0.103.2
h11==0.14.0
idna==3.4
passlib==1.7.4
pyasn1==0.5.0
pydantic==2.4.2
pydantic_core==2.10.1
pymongo==4.5.0
python-jose==3.3.0
python-multipart==0.0.6
rsa==4.9
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.27.0
typing_extensions==4.8.0
tzdata==2023.3
uvicorn==0.23.2
